Fix Environment fields & move test files into right package
